+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ** 1.1.9-dev **
 
   - added `addItem` and `removeItem` functions to validate dynamically created fields
+  - added addConstraint, updateConstraint and removeConstraint API to dynamically
+    update fields constraints
 
 ** 1.1.8 **